Move component to views folder
