+++ packages/client-web/src/routes/Home/index.tsx
@@ -5,7 +5,7 @@ import { HorizontalLoader } from "@quoll/ui-components";
 
 import { useTimelineViewModel } from "../../modules/timeline/view-model";
 import { useDateViewModelModel } from "../../modules/date/view-model";
-import DatePicker from "../../components/DatePicker";
+import DatePicker from "../../modules/date/views/DatePicker";
 import Timeline from "../../modules/timeline/views/Timeline";
 import Map from "../../components/Map";
 import { makePolylineConfigs, makeInfoWindowOptions } from "./mapUtils";
